Remove fillTree() add inOrder() and newNode()
Removed fillTree function and added inOrder function that uses recursion
to print the inOrder traversal of the binary tree that is passed into
the function. newNode function sets the int value passed into the
function to the node and sets the left and right values to NULL. Also
added a visual representation of the binary tree using ascii art in a
comment.

The fillTree function was removed as it was not useful to have a binary
tree filled with random values for testing the binary tree traversal
algorithms. The newNode function was added to enable manual creation of
nodes with specific values. The inOrder function was added to progress
through the tree in order from left to root to right and print the value
in each node.