Add treeHeight() to find height of the binary tree
Add treeHeight, a function that takes in a binary tree and returns an
int value that is the height of the binary tree. Add max function to be
used in treeHeight function to find the max height between the left
and right side of the tree. The max function takes in 2 int values and
returns the value that is greater.

The treeHeight function allows the user to easily find the height of the
binary tree.